Specify molecule content for all 11 bands
By specifying the molecule content for
a spectral band, its clirad `param` dictionary
can be looked up.  Using this `param`, the
analysis directory of the clirad run corresponding
to it is known.  Inside this directory, the file
'param.py' contains the `param` for the clirad run,
 as well as the `param` for the lblnew run that it
was compared against.  The fluxes and cooling rate
for the lblnew run can therefore be retrieved from
its analysis directory.