Define way to load single-run clirad results for atm
Before the results from separate clirad-lw runs are
collected together to get the total result for a typical
atmosphere.  But, in practice, by specifying all the bands
and molecule concs for the atmosphere at input, clirad-lw
only has to be run once to produce the results.

Here, such an input if defined, and the previous function
of loading clirad results has been modified such that
it can output total result for the atmosphere from
multi-run results or from single-run results.