Results after taking out sfcflux()
We don't call sfcflux() because we suspect that
it is buggy, and we replace its output by manually
setting values for blayer(np+1) in each spectral band.

However, we have only done this for a surface temperature
of 294 K, so the results for saw and trp profiles are no
longer valid.  Therefore, we will only used mls profile
for testing.