Create notebook to compare cliradlw with lblnew-bestfit
This notebook is for comparing clirad-lw with lblnew-bestfit.
In this notebook, one or more pairs of input parameter
dictionaries, one clirad-lw's and the other lblnew-bestfit's, can be
specified.  For each pair, a separate notebook will be
created, in which the cooling rate and fluxes are compared
between clirad-lw and lblnew-bestfit.  This notebook
is saved at a recoverable file path.