Create notebook to compare cliradlw with lblnew-overlap
This notebook is for comparing clirad-lw with lblnew-overlap.
In this notebook, one or more pairs of input parameter
dictionaries, one clirad-lw's and the other lblnew-overlap's, can be
specified.  For each pair, a separate notebook will be
created, in which the cooling rate and fluxes are compared
between clirad-lw and lblnew-overlap.  This notebook
is saved at a recoverable file path.