Figures for the final draft
These figures have been checked by Prof Chou and are included
in the final draft that was sent to Wei Liang to be submitted
to a journal.

Note that the climatools versions in this notebook is not the
latest version, because the latest version was having some trouble
retrieving the correct dataset from mongodb.